Backend/requirements.txt
appdirs
argh
asn1crypto
astropy
async-generator
autopep8
backports.shutil-get-terminal-size
backtesting
bcrypt
bitarray
bkcharts
boto
bottleneck
brotlipy
ccxt
colorama
contextlib2
cython
cytoolz
distributed
django
fastapi-users
fastcache
flake8
gevent
glob2
gmpy2
html5lib
httptools
install
jdcal
jose
jupyter
jupyter-dash
jupyter-packaging
jupyterlab
mkl-fft
mkl-random
mkl-service
mock
more-itertools
motor
multipledispatch
nltk
nose
numba
olefile
openpyxl
pandas-datareader
path
pathlib2
pep8
pip-chill
pip-upgrader
plotly-express
ply
pycurl
pydocstyle
pyls-black
pyodbc
pyopenssl
pysocks
pytest
python-binance
python-dotenv
python-jose
rope
routes
scikit-image
seaborn
setuptools
simplegeneric
singledispatch
sip
sklearn
sortedcollections
sphinxcontrib-websupport
spyder
sqlalchemy
stack-data
sympy
tables
tensorflow
typed-ast
unicodecsv
uvicorn
uvloop
watchgod
xlrd
xlsxwriter
xlwt
xmltodict
yapf
